AgentLab/requirements.txt
# BrowserGym and AgentLab Dependencies
# Install with: pip install -r requirements.txt

# Core BrowserGym packages
browsergym>=0.14.0
browsergym-core>=0.14.0
browsergym-experiments>=0.14.0

# AgentLab
agentlab>=0.2.0

# LLM providers (choose based on your preference)
openai>=1.0.0
anthropic>=0.18.0

# Playwright for browser automation
playwright>=1.40.0

# Utility packages
python-dotenv>=1.0.0
pydantic>=2.0.0
gymnasium>=0.29.0

# For data analysis and visualization
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0

# Optional: for async operations
aiohttp>=3.9.0
asyncio-throttle>=1.0.0




